requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.7.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coverage==4.5.4
decorator==4.4.1
Django==2.2
django-debug-toolbar==2.1
djangorestframework==3.11.0
django-rest-swagger==2.2.0
entrypoints==0.3
flake8==3.7.9
flower==0.9.3
idna==2.8
importlib-metadata==1.3.0
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
kombu==4.6.7
mccabe==0.6.1
more-itertools==8.0.2
parso==0.5.2
pickleshare==0.7.5
prompt-toolkit==3.0.2
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pytz==2019.3
redis==3.3.11
requests==2.22.0
sentry-sdk==0.13.5
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0